+++ src/Pages/Assets/Assets.js
@@ -1,10 +1,13 @@
 import React from 'react';
 import { Section, Wrapper, Header } from 'Components';
-import { AssetsList } from './Components';
+import { AssetsList, AssetsAUM } from './Components';
 
 const Assets = () => (
   <Wrapper>
     <Header title="All Assets" />
+    <Section header>
+      <AssetsAUM />
+    </Section>
     <Section header>
       <AssetsList />
     </Section>